refactor: move subcomponent to new file
