+++ src/views/components/PageSelection/PageSelection.tsx
@@ -6,9 +6,8 @@ import {
 	PageSelectionViewModel
 } from '@src/view-models/PageSelectionViewModel.ts'
 import {
-	GamePageListViewModel
-} from '@src/view-models/GamePageListViewModel.ts'
-import { Dispatch, SetStateAction } from 'react'
+	CustomPageSelection
+} from '@views/components/CustomPageSelection/CustomPageSelection.tsx'
 
 
 const viewModel = new PageSelectionViewModel()
@@ -47,30 +46,5 @@ export function PageSelection(props: IPageSelection) {
 								 setCurrentPage={setCurrentPage}
 								 pageIndices={pageIndices}/>
 		</aside>
-	)
-}
-
-export interface ICustomPageSelection {
-	parentViewModel: GamePageListViewModel
-	setCurrentPage: Dispatch<SetStateAction<number>>
-	pageIndices: number[]
-}
-
-function CustomPageSelection(props: ICustomPageSelection) {
-	const { parentViewModel, setCurrentPage, pageIndices } = props
-
-	return (
-		<form className={style.CustomPageSelection} onSubmit={(e) => {
-			parentViewModel.changeCustomPage(e,
-											 setCurrentPage,
-											 pageIndices.length)
-		}}>
-			<p>
-				<label htmlFor="custom-page">enter page: </label>
-				<input type="tel" name="custom-page" id="custom-page"/>
-			</p>
-
-			<button type="submit">save</button>
-		</form>
 	)
 }
